chore(dsl): Split too large test class
