+++ kotlin-dsl/dsl/src/test/kotlin/com/chutneytesting/kotlin/synchronize/server/ChutneyServerServiceImplTest.kt
@@ -1,0 +1,57 @@
+/*
+ *  Copyright 2017-2023 Enedis
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package com.chutneytesting.kotlin.synchronize.server
+
+import com.chutneytesting.kolin.util.ChutneyServerInfoClearProperties
+import com.chutneytesting.kotlin.util.ChutneyServerInfo
+import com.chutneytesting.kotlin.util.HttpsClientTest
+import com.fasterxml.jackson.databind.JsonNode
+import com.fasterxml.jackson.databind.ObjectMapper
+import com.github.tomakehurst.wiremock.core.WireMockConfiguration
+import com.github.tomakehurst.wiremock.junit5.WireMockExtension
+import com.github.tomakehurst.wiremock.stubbing.ServeEvent
+import org.junit.jupiter.api.extension.RegisterExtension
+
+@ChutneyServerInfoClearProperties
+abstract class ChutneyServerServiceImplTest {
+
+    private val keystorePath: String = HttpsClientTest::class.java.getResource("/security/server.jks")!!.path
+
+    @RegisterExtension
+    val wireMockServer: WireMockExtension = WireMockExtension.newInstance()
+        .options(
+            WireMockConfiguration.wireMockConfig()
+            .dynamicHttpsPort()
+            .httpDisabled(true)
+            .keystorePath(keystorePath)
+            .keystorePassword("server")
+            .keyManagerPassword("server")
+        )
+        .configureStaticDsl(false)
+        .build()
+
+    fun serverEventRequestBodyAsJson(event: ServeEvent): JsonNode {
+        return ObjectMapper().readTree(event.request.bodyAsString)
+    }
+
+    fun buildServerInfo() = ChutneyServerInfo(
+        wireMockServer.baseUrl(),
+        "user",
+        "password"
+    )
+}
